refactor: 디렉토리 구조 리팩토링
매핑 역할의 도메인을 상위 도메인의 하위도메인으로서 디렉토리 구조를 위치할 수 있도록 함.
DDD에 가까운 디렉토리 구조로 변경
