+++ src/modules/tasks/dtos/get-task.dto.ts
@@ -1,10 +1,10 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Status } from '../../../common/enums/status.enum';
-import { ManagerResponseDto } from '../../mappings/managers/dtos/create-manager-dto';
-import { TaskFileResponseDto } from '../../mappings/task-files/dtos/create-task-files.dto';
+import { ManagerResponseDto } from './create-manager-dto';
+import { TaskFileResponseDto } from '../task-files/dtos/create-task-files.dto';
 import { Task } from '../entities/tasks.entity';
-import { Manager } from '../../mappings/managers/managers.entity';
-import { TaskFile } from '../../mappings/task-files/task-files.entity';
+import { Manager } from '../entities/managers.entity';
+import { TaskFile } from '../task-files/task-files.entity';
 
 export class GetTaskResponseDto {
     @ApiProperty({
