+++ src/modules/tasks/dtos/update-task.dto.ts
@@ -1,10 +1,10 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Status } from '../../../common/enums/status.enum';
 import { IsOptional, IsNotEmpty, IsEnum, IsArray, IsNumber } from 'class-validator';
-import { ManagerResponseDto } from '../../mappings/managers/dtos/create-manager-dto';
+import { ManagerResponseDto } from './create-manager-dto';
 import { Type } from 'class-transformer';
 import { Task } from '../entities/tasks.entity';
-import { Manager } from '../../mappings/managers/managers.entity';
+import { Manager } from '../entities/managers.entity';
 export class UpdateTaskRequestDto {
     @ApiProperty({
         example: 'api명세서 작성',
