+++ src/modules/tasks/entities/tasks.entity.ts
@@ -2,8 +2,8 @@ import { Status } from 'src/common/enums/status.enum';
 import { Column, Entity, ManyToOne, OneToMany } from 'typeorm';
 import { Comment } from '../../comments/entities/comments.entity';
 import { BaseEntity } from 'src/common/entities/base.entity';
-import { Manager } from '../../mappings/managers/managers.entity';
-import { TaskFile } from '../../mappings/task-files/task-files.entity';
+import { Manager } from './managers.entity';
+import { TaskFile } from '../task-files/task-files.entity';
 import { Step } from '../../steps/entities/steps.entity';
 
 @Entity()
