+++ src/modules/tasks/tasks.controller.ts
@@ -43,7 +43,7 @@ import {
 import { Status } from '../../common/enums/status.enum';
 import { Transactional } from 'src/common/decorators/transaction.decorator';
 import { TransactionalRequest } from 'src/common/decorators/transaction.decorator';
-import { CreateTaskFileResponseDto } from '../mappings/task-files/dtos/create-task-files.dto';
+import { CreateTaskFileResponseDto } from './task-files/dtos/create-task-files.dto';
 import { DeleteTaskResponseDto } from './dtos/delete-task.dto';
 import { GetCommentResponseDto } from '../comments/dto/get-comment.dto';
 import { ProjectDashBoardDTO } from './dtos/user-task.dto';
