+++ src/modules/tasks/tasks.module.ts
@@ -6,8 +6,8 @@ import { Task } from './entities/tasks.entity';
 import { Step } from '../steps/entities/steps.entity';
 import { UserProject } from '../projects/user-projects/entities/user-projects.entity';
 import { ConfigModule } from '@nestjs/config';
-import { Manager } from '../mappings/managers/managers.entity';
-import { TaskFile } from '../mappings/task-files/task-files.entity';
+import { Manager } from './entities/managers.entity';
+import { TaskFile } from './task-files/task-files.entity';
 import { UploadService } from '../../infra/upload/upload.service';
 import { Project } from '../projects/entities/projects.entity';
 import { Comment } from '../comments/entities/comments.entity';
@@ -15,10 +15,10 @@ import { UsersModule } from '../users/users.module';
 import { TaskRepository } from './repositories/task.repository';
 import { CommentRepository } from '../comments/repositories/comments.repository';
 import { StepRepository } from '../steps/repositories/step.repository';
-import { TaskFileRepository } from '../mappings/task-files/repositories/task-file.repository';
-import { ManagerRepository } from '../mappings/managers/repositories/manager.repository';
+import { TaskFileRepository } from './task-files/repositories/task-file.repository';
 import { ProjectsModule } from '../projects/projects.module';
 import { UserProjectModule } from '../projects/user-projects/user-project.module';
+import { ManagerRepository } from './repositories/manager.repository';
 @Module({
     imports: [
         TypeOrmModule.forFeature([Task, Step, UserProject, Manager, TaskFile, Project, Comment]),
