+++ src/modules/users/entities/users.entity.ts
@@ -3,9 +3,9 @@ import { Comment } from '../../comments/entities/comments.entity';
 import { BaseEntity } from 'src/common/entities/base.entity';
 import { PersonalRecall } from '../../personal-recalls/entities/personal-recalls.entity';
 import { UserProject } from '../../projects/user-projects/entities/user-projects.entity';
-import { Manager } from '../../mappings/managers/managers.entity';
-import { TaskFile } from '../../mappings/task-files/task-files.entity';
-import { ProjectFile } from '../../mappings/project-files/project-files.entity';
+import { Manager } from '../../tasks/entities/managers.entity';
+import { TaskFile } from '../../tasks/task-files/task-files.entity';
+import { ProjectFile } from '../../projects/entities/project-files.entity';
 import { Writer } from '../../plans/entities/writers.entity';
 import { Attendee } from '../../plans/entities/attendees.entity';
 import { MasterPortfolio } from '../../master-portfolios/entities/master-portfolios.entity';
